refactor(notebook-cloud): isolate hosted live cell renderer
